+++ src/characteristic/positionnable/Moveable.java
@@ -4,5 +4,5 @@ import javafx.geometry.Point3D;
 
 public interface Moveable extends Positionnable{
 	public void moveTo(Point3D nextPosition);
-	public Point3D computeNextPosition(double secondsPassed);
+	public Point3D computeNextPosition(float secondsPassed);
 }
